[connector] Rename rf connector to be aligned with the connector usages + add documentation
