+++ external-import/recorded-future/docker-compose.yml
@@ -1,30 +1,30 @@
 version: '3'
 services:
   connector-recordedfuture-notes:
-    image: opencti/connector-recordedfuture-notes:5.12.32
+    image: opencti/connector-recorded-future:5.12.29
     environment:
       - OPENCTI_URL=http://opencti:8080
       - OPENCTI_TOKEN=CHANGME
       - CONNECTOR_ID=CHANGEME
-      - CONNECTOR_NAME=Recorded Future Analyst Notes
+      - CONNECTOR_TYPE=EXTERNAL_IMPORT
+      - CONNECTOR_NAME=Recorded Future
       - CONNECTOR_SCOPE=ipv4-addr,ipv6-addr,vulnerability,domain,url,StixFile
-      - CONNECTOR_CONFIDENCE_LEVEL=5
       - CONNECTOR_LOG_LEVEL=error
       - RECORDED_FUTURE_TOKEN=CHANGEME #required
       - RECORDED_FUTURE_INITIAL_LOOKBACK=240 #required
       - RECORDED_FUTURE_INTERVAL=240 #required
-      - RECORDED_FUTURE_TLP=White # required
-      - RECORDED_FUTURE_PULL_SIGNATURES=False #optional, can remove
-      - RECORDED_FUTURE_INSIKT_ONLY=True #optional, can remove
       - RECORDED_FUTURE_TOPIC=ChangeMe #optional, can remove
+      - RECORDED_FUTURE_TLP=White # required
+      - RECORDED_FUTURE_INSIKT_ONLY=True #optional, can remove
+      - RECORDED_FUTURE_PULL_SIGNATURES=False #optional, can remove
       - RECORDED_FUTURE_PERSON_TO_TA=False #optional, can remove
       - RECORDED_FUTURE_TA_TO_INTRUSION_SET=False #optional, can remove
       - RECORDED_FUTURE_RISK_AS_SCORE=True #optional, can remove
       - RECORDED_FUTURE_RISK_THRESHOLD=60 #optional, can remove
       - RECORDED_FUTURE_PULL_RISK_LIST=False #optional, can remove
+      - RECORDED_FUTURE_RISK_LIST_INTERVAL=6 #required, in hours
       - RECORDED_FUTURE_RISK_LIST_THRESHOLD=70 #optional, can remove
       - RECORDED_FUTURE_RISKLIST_RELATED_ENTITIES='Malware,Hash,URL,Threat Actor,MitreAttackIdentifier' #required if RECORDED_FUTURE_PULL_RISK_LIST is True, possible values: Malware,Hash,URL,Threat Actor,MitreAttackIdentifier
-      - RECORDED_FUTURE_RISK_LIST_INTERVAL=6 #required, in hours
       - RECORDED_FUTURE_PULL_THREAT_MAPS=False #optional, can remove
       - RECORDED_FUTURE_THREAT_MAPS_INTERVAL=24 #required, in hours
     restart: always
