+++ external-import/recorded-future/src/config.yml.sample
@@ -5,29 +5,28 @@ opencti:
 connector:
   id: 'ChangeMe'
   type: 'EXTERNAL_IMPORT'
-  name: 'Recorded Future Analyst Notes' # Required
+  name: 'Recorded Future' # Required
   scope: 'ipv4-addr,ipv6-addr,vulnerability,domain,url,file-sha256,file-md5,file-sha1'
-  confidence_level: 5
   update_existing_data: false
   log_level: 'info'
 
 rf-notes:
   token: 'ChangeMe' # Required
   initial_lookback: 240 # Required, in hours
-  interval: 24 # Required, in hours
-  risk_list_interval: 6 # Required, in hours
   TLP: white
+  interval: 24 # Required, in hours
+  topic: 'VTrvnW,g1KBGl,ZjnoP0,aDKkpk,TXSFt5,UrMRnT,TXSFt3' # optional / Corresponds to Yara Rule, Sigma Rule, Snort Rule, TTP Instance, Validated Intelligence Event, Informational, Threat Lead
+  insikt_only: True # optional
+  pull_signatures: False # Pull Yara/Snort/Sigma rules into OpenCTI. optional
+  person_to_TA: False # optional
+  TA_to_intrusion_set: False # optional
+  risk_as_score: True # optional
+  risk_threshold: 60 # optional
   pull_risk_list: False
+  risk_list_interval: 6 # Required, in hours
   risk_list_threshold: 70 # optional
   # if pull_risk_list is true, risklist_related_entities is required.
   # Available choices: Malware,Hash,URL,Threat Actor,MitreAttackIdentifier
-  risklist_related_entities: 'Malware,Threat Actor,MitreAttackIdentifier'
-  pull_signatures: False # Pull Yara/Snort/Sigma rules into OpenCTI. optional
-  insikt_only: True # optional
-  topic: 'VTrvnW,g1KBGl,ZjnoP0,aDKkpk,TXSFt5,UrMRnT,TXSFt3' # optional / Corresponds to Yara Rule, Sigma Rule, Snort Rule, TTP Instance, Validated Intelligence Event, Informational, Threat Lead
-  person_to_TA: False # optional
-  TA_to_intrusion_set: False # optional
-  risk_as_score: False # optional
-  risk_threshold: 60 # optional
+  risklist_related_entities: 'Malware,Threat Actor,MitreAttackIdentifier',
   pull_threat_maps: False # optional - Pull Threat Actors and Malware maps
   threat_maps_interval: 24 # Required in hours
