+++ external-import/recorded-future/src/rflib/__init__.py
@@ -13,7 +13,7 @@
 from ._version import __version__ as APP_VERSION
 from .constants import RISK_LIST_TYPE_MAPPER
 from .rf_client import RFClient
-from .rf_notes_to_stix2 import (
+from .rf_to_stix2 import (
     TTP,
     URL,
     Domain,
